Created a generic input file to store all variables and also set some default values to allow simple testing. Have updated the way time and temperature units are stored to simplify converstion steps. The crystal and time teperature information can then be used in the Monte Carlo simulation struct. Work can now focus on the actual monte carlo simulation.

+++ common/src/numeric.rs
@@ -15,8 +15,7 @@ use rand::Rng;
 use ndarray::{Array, Array1, ArrayBase, ArrayD, Data, Dimension, Zip};
 
 /// Default floating-point precision used for model parameters and state.
-pub type Float = f32;
-// pub type Float = f64;
+pub type Float = f64;
 /// Higher precision used for rates, lifetimes, and time/temperature profiles.
 pub type TimeFloat = f64;
